Create Natural Cubic Spline Approximations for Continuous Time

Accepts or constructs a natural cubic spline basis for continuous time and yields a matrix of approximations for time according to that basis.

Usage

time_spline(
  time,
  df = NULL,
  ...,
  basis = time_spline_basis(time, df = df, ...)
)

Arguments

time

A numeric vector of values.

df, ...

Only used if basis is left as the default. Passed to time_spline_basis() (which passes all arguments to splines::ns()) to calculate the spline basis.

basis

Spline basis for which to create approximations of time. Defaults to time_spline_basis(time, df = df, ...).

Value

Matrix with the same dimensions as basis. Contains basis as an attribute.

Details

time_spline() is primarily useful because it can use one step to create the spline basis from time and then re-input time into the spline basis to obtain the spline approximations. Alternatively, it can calculate predictions from a basis supplied to the basis argument.

Examples

time_spline(Theoph$Time, df = 3)
#>                   1          2           3
#>   [1,]  0.000000000 0.00000000  0.00000000
#>   [2,] -0.031886120 0.07479893 -0.04282603
#>   [3,] -0.070807627 0.16805702 -0.09622083
#>   [4,] -0.125964695 0.31293414 -0.17916995
#>   [5,] -0.169412675 0.48806315 -0.27927835
#>   [6,] -0.115368472 0.64343847 -0.36471168
#>   [7,] -0.005439624 0.64376443 -0.35709820
#>   [8,]  0.187939184 0.56594061 -0.28713439
#>   [9,]  0.336223856 0.48518463 -0.19148429
#>  [10,]  0.432921151 0.40296875 -0.02765524
#>  [11,] -0.140818100 0.37708366  0.76373230
#>  [12,]  0.000000000 0.00000000  0.00000000
#>  [13,] -0.034401475 0.08073620 -0.04622541
#>  [14,] -0.064955064 0.15378594 -0.08804994
#>  [15,] -0.115680778 0.28362245 -0.16238767
#>  [16,] -0.167320051 0.47227657 -0.27028435
#>  [17,] -0.135633825 0.63136451 -0.35893139
#>  [18,] -0.013265008 0.64554696 -0.35878460
#>  [19,]  0.187939184 0.56594061 -0.28713439
#>  [20,]  0.333419797 0.48691072 -0.19397616
#>  [21,]  0.431739620 0.40536423 -0.03443175
#>  [22,] -0.135520210 0.37659567  0.75892036
#>  [23,]  0.000000000 0.00000000  0.00000000
#>  [24,] -0.034401475 0.08073620 -0.04622541
#>  [25,] -0.071965460 0.17089459 -0.09784548
#>  [26,] -0.117472923 0.28861038 -0.16524350
#>  [27,] -0.169412675 0.48806315 -0.27927835
#>  [28,] -0.128469570 0.63656288 -0.36151519
#>  [29,] -0.007403778 0.64422798 -0.35753255
#>  [30,]  0.191599626 0.56411346 -0.28534422
#>  [31,]  0.333419797 0.48691072 -0.19397616
#>  [32,]  0.433186017 0.40237949 -0.02595674
#>  [33,] -0.125686467 0.37569104  0.74998465
#>  [34,]  0.000000000 0.00000000  0.00000000
#>  [35,] -0.044374820 0.10436966 -0.05975672
#>  [36,] -0.074267838 0.17655229 -0.10108478
#>  [37,] -0.121816839 0.30090106 -0.17228051
#>  [38,] -0.170988539 0.50443930 -0.28859376
#>  [39,] -0.135633825 0.63136451 -0.35893139
#>  [40,] -0.013265008 0.64554696 -0.35878460
#>  [41,]  0.187019320 0.56639890 -0.28758125
#>  [42,]  0.334546483 0.48621869 -0.19298014
#>  [43,]  0.431523624 0.40576949 -0.03555843
#>  [44,] -0.162020769 0.37903913  0.78298164
#>  [45,]  0.000000000 0.00000000  0.00000000
#>  [46,] -0.038158943 0.08962167 -0.05131277
#>  [47,] -0.064955064 0.15378594 -0.08804994
#>  [48,] -0.115680778 0.28362245 -0.16238767
#>  [49,] -0.169412675 0.48806315 -0.27927835
#>  [50,] -0.135633825 0.63136451 -0.35893139
#>  [51,] -0.013265008 0.64554696 -0.35878460
#>  [52,]  0.187019320 0.56639890 -0.28758125
#>  [53,]  0.338985800 0.48347180 -0.18898637
#>  [54,]  0.431739620 0.40536423 -0.03443175
#>  [55,] -0.139304250 0.37694418  0.76235743
#>  [56,]  0.000000000 0.00000000  0.00000000
#>  [57,] -0.034401475 0.08073620 -0.04622541
#>  [58,] -0.071965460 0.17089459 -0.09784548
#>  [59,] -0.128360092 0.32003121 -0.18323322
#>  [60,] -0.169587065 0.48959437 -0.28015006
#>  [61,] -0.131520539 0.63449736 -0.36050068
#>  [62,] -0.015207904 0.64596224 -0.35918470
#>  [63,]  0.185173915 0.56731723 -0.28847414
#>  [64,]  0.345443447 0.47941486 -0.18296684
#>  [65,]  0.432737810 0.40336372 -0.02878661
#>  [66,] -0.101522574 0.37347754  0.72799512
#>  [67,]  0.000000000 0.00000000  0.00000000
#>  [68,] -0.031886120 0.07479893 -0.04282603
#>  [69,] -0.062585869 0.14804054 -0.08476042
#>  [70,] -0.117472923 0.28861038 -0.16524350
#>  [71,] -0.169412675 0.48806315 -0.27927835
#>  [72,] -0.136774489 0.63041696 -0.35845070
#>  [73,] -0.015207904 0.64596224 -0.35918470
#>  [74,]  0.183321049 0.56823778 -0.28936584
#>  [75,]  0.333419797 0.48691072 -0.19397616
#>  [76,]  0.432255723 0.40435862 -0.03161162
#>  [77,] -0.129467769 0.37603869  0.75342133
#>  [78,]  0.000000000 0.00000000  0.00000000
#>  [79,] -0.031886120 0.07479893 -0.04282603
#>  [80,] -0.064955064 0.15378594 -0.08804994
#>  [81,] -0.113858263 0.27859466 -0.15950901
#>  [82,] -0.169412675 0.48806315 -0.27927835
#>  [83,] -0.133893864 0.63274194 -0.35962540
#>  [84,] -0.010340372 0.64490104 -0.35816820
#>  [85,]  0.198829456 0.56048783 -0.28175081
#>  [86,]  0.337333680 0.48449791 -0.19048585
#>  [87,]  0.432737810 0.40336372 -0.02878661
#>  [88,] -0.121906461 0.37534379  0.74654816
#>  [89,]  0.000000000 0.00000000  0.00000000
#>  [90,] -0.038158943 0.08962167 -0.05131277
#>  [91,] -0.077687049 0.18499373 -0.10591792
#>  [92,] -0.120102880 0.29601578 -0.16948345
#>  [93,] -0.169412675 0.48806315 -0.27927835
#>  [94,] -0.133893864 0.63274194 -0.35962540
#>  [95,] -0.013265008 0.64554696 -0.35878460
#>  [96,]  0.200617994 0.55958737 -0.28084975
#>  [97,]  0.321779716 0.49394857 -0.20388285
#>  [98,]  0.426370035 0.41379984 -0.05681476
#>  [99,] -0.145360341 0.37750232  0.76785698
#> [100,]  0.000000000 0.00000000  0.00000000
#> [101,] -0.046843101 0.11024514 -0.06312071
#> [102,] -0.093041003 0.22359645 -0.12801985
#> [103,] -0.117472923 0.28861038 -0.16524350
#> [104,] -0.169917035 0.49263118 -0.28187848
#> [105,] -0.132714772 0.63363118 -0.36007016
#> [106,] -0.010340372 0.64490104 -0.35816820
#> [107,]  0.192509988 0.56365816 -0.28489600
#> [108,]  0.353680730 0.47412303 -0.17488729
#> [109,]  0.432737810 0.40336372 -0.02878661
#> [110,] -0.090224045 0.37244888  0.71769159
#> [111,]  0.000000000 0.00000000  0.00000000
#> [112,] -0.031886120 0.07479893 -0.04282603
#> [113,] -0.062585869 0.14804054 -0.08476042
#> [114,] -0.113858263 0.27859466 -0.15950901
#> [115,] -0.168652039 0.48185238 -0.27574137
#> [116,] -0.129701084 0.63575371 -0.36111993
#> [117,] -0.013265008 0.64554696 -0.35878460
#> [118,]  0.187939184 0.56594061 -0.28713439
#> [119,]  0.335107293 0.48587347 -0.19248177
#> [120,]  0.432921151 0.40296875 -0.02765524
#> [121,] -0.118883480 0.37506632  0.74379911
#> [122,]  0.000000000 0.00000000  0.00000000
#> [123,] -0.031886120 0.07479893 -0.04282603
#> [124,] -0.062585869 0.14804054 -0.08476042
#> [125,] -0.115680778 0.28362245 -0.16238767
#> [126,] -0.169045005 0.48497498 -0.27751992
#> [127,] -0.134477691 0.63228862 -0.35939765
#> [128,] -0.008383944 0.64445532 -0.35774656
#> [129,]  0.191599626 0.56411346 -0.28534422
#> [130,]  0.335107293 0.48587347 -0.19248177
#> [131,]  0.432255723 0.40435862 -0.03161162
#> [132,] -0.124174302 0.37555209  0.74861003
#> attr(,"degree")
#> [1] 3
#> attr(,"knots")
#> [1] 1.040000 7.023333
#> attr(,"Boundary.knots")
#> [1]  0.00 24.65
#> attr(,"intercept")
#> [1] FALSE
#> attr(,"class")
#> [1] "splinetrials_time_spline" "ns"                      
#> [3] "basis"                    "matrix"                  
#> attr(,"basis")
#>                   1          2           3
#>   [1,]  0.000000000 0.00000000  0.00000000
#>   [2,] -0.031886120 0.07479893 -0.04282603
#>   [3,] -0.070807627 0.16805702 -0.09622083
#>   [4,] -0.125964695 0.31293414 -0.17916995
#>   [5,] -0.169412675 0.48806315 -0.27927835
#>   [6,] -0.115368472 0.64343847 -0.36471168
#>   [7,] -0.005439624 0.64376443 -0.35709820
#>   [8,]  0.187939184 0.56594061 -0.28713439
#>   [9,]  0.336223856 0.48518463 -0.19148429
#>  [10,]  0.432921151 0.40296875 -0.02765524
#>  [11,] -0.140818100 0.37708366  0.76373230
#>  [12,]  0.000000000 0.00000000  0.00000000
#>  [13,] -0.034401475 0.08073620 -0.04622541
#>  [14,] -0.064955064 0.15378594 -0.08804994
#>  [15,] -0.115680778 0.28362245 -0.16238767
#>  [16,] -0.167320051 0.47227657 -0.27028435
#>  [17,] -0.135633825 0.63136451 -0.35893139
#>  [18,] -0.013265008 0.64554696 -0.35878460
#>  [19,]  0.187939184 0.56594061 -0.28713439
#>  [20,]  0.333419797 0.48691072 -0.19397616
#>  [21,]  0.431739620 0.40536423 -0.03443175
#>  [22,] -0.135520210 0.37659567  0.75892036
#>  [23,]  0.000000000 0.00000000  0.00000000
#>  [24,] -0.034401475 0.08073620 -0.04622541
#>  [25,] -0.071965460 0.17089459 -0.09784548
#>  [26,] -0.117472923 0.28861038 -0.16524350
#>  [27,] -0.169412675 0.48806315 -0.27927835
#>  [28,] -0.128469570 0.63656288 -0.36151519
#>  [29,] -0.007403778 0.64422798 -0.35753255
#>  [30,]  0.191599626 0.56411346 -0.28534422
#>  [31,]  0.333419797 0.48691072 -0.19397616
#>  [32,]  0.433186017 0.40237949 -0.02595674
#>  [33,] -0.125686467 0.37569104  0.74998465
#>  [34,]  0.000000000 0.00000000  0.00000000
#>  [35,] -0.044374820 0.10436966 -0.05975672
#>  [36,] -0.074267838 0.17655229 -0.10108478
#>  [37,] -0.121816839 0.30090106 -0.17228051
#>  [38,] -0.170988539 0.50443930 -0.28859376
#>  [39,] -0.135633825 0.63136451 -0.35893139
#>  [40,] -0.013265008 0.64554696 -0.35878460
#>  [41,]  0.187019320 0.56639890 -0.28758125
#>  [42,]  0.334546483 0.48621869 -0.19298014
#>  [43,]  0.431523624 0.40576949 -0.03555843
#>  [44,] -0.162020769 0.37903913  0.78298164
#>  [45,]  0.000000000 0.00000000  0.00000000
#>  [46,] -0.038158943 0.08962167 -0.05131277
#>  [47,] -0.064955064 0.15378594 -0.08804994
#>  [48,] -0.115680778 0.28362245 -0.16238767
#>  [49,] -0.169412675 0.48806315 -0.27927835
#>  [50,] -0.135633825 0.63136451 -0.35893139
#>  [51,] -0.013265008 0.64554696 -0.35878460
#>  [52,]  0.187019320 0.56639890 -0.28758125
#>  [53,]  0.338985800 0.48347180 -0.18898637
#>  [54,]  0.431739620 0.40536423 -0.03443175
#>  [55,] -0.139304250 0.37694418  0.76235743
#>  [56,]  0.000000000 0.00000000  0.00000000
#>  [57,] -0.034401475 0.08073620 -0.04622541
#>  [58,] -0.071965460 0.17089459 -0.09784548
#>  [59,] -0.128360092 0.32003121 -0.18323322
#>  [60,] -0.169587065 0.48959437 -0.28015006
#>  [61,] -0.131520539 0.63449736 -0.36050068
#>  [62,] -0.015207904 0.64596224 -0.35918470
#>  [63,]  0.185173915 0.56731723 -0.28847414
#>  [64,]  0.345443447 0.47941486 -0.18296684
#>  [65,]  0.432737810 0.40336372 -0.02878661
#>  [66,] -0.101522574 0.37347754  0.72799512
#>  [67,]  0.000000000 0.00000000  0.00000000
#>  [68,] -0.031886120 0.07479893 -0.04282603
#>  [69,] -0.062585869 0.14804054 -0.08476042
#>  [70,] -0.117472923 0.28861038 -0.16524350
#>  [71,] -0.169412675 0.48806315 -0.27927835
#>  [72,] -0.136774489 0.63041696 -0.35845070
#>  [73,] -0.015207904 0.64596224 -0.35918470
#>  [74,]  0.183321049 0.56823778 -0.28936584
#>  [75,]  0.333419797 0.48691072 -0.19397616
#>  [76,]  0.432255723 0.40435862 -0.03161162
#>  [77,] -0.129467769 0.37603869  0.75342133
#>  [78,]  0.000000000 0.00000000  0.00000000
#>  [79,] -0.031886120 0.07479893 -0.04282603
#>  [80,] -0.064955064 0.15378594 -0.08804994
#>  [81,] -0.113858263 0.27859466 -0.15950901
#>  [82,] -0.169412675 0.48806315 -0.27927835
#>  [83,] -0.133893864 0.63274194 -0.35962540
#>  [84,] -0.010340372 0.64490104 -0.35816820
#>  [85,]  0.198829456 0.56048783 -0.28175081
#>  [86,]  0.337333680 0.48449791 -0.19048585
#>  [87,]  0.432737810 0.40336372 -0.02878661
#>  [88,] -0.121906461 0.37534379  0.74654816
#>  [89,]  0.000000000 0.00000000  0.00000000
#>  [90,] -0.038158943 0.08962167 -0.05131277
#>  [91,] -0.077687049 0.18499373 -0.10591792
#>  [92,] -0.120102880 0.29601578 -0.16948345
#>  [93,] -0.169412675 0.48806315 -0.27927835
#>  [94,] -0.133893864 0.63274194 -0.35962540
#>  [95,] -0.013265008 0.64554696 -0.35878460
#>  [96,]  0.200617994 0.55958737 -0.28084975
#>  [97,]  0.321779716 0.49394857 -0.20388285
#>  [98,]  0.426370035 0.41379984 -0.05681476
#>  [99,] -0.145360341 0.37750232  0.76785698
#> [100,]  0.000000000 0.00000000  0.00000000
#> [101,] -0.046843101 0.11024514 -0.06312071
#> [102,] -0.093041003 0.22359645 -0.12801985
#> [103,] -0.117472923 0.28861038 -0.16524350
#> [104,] -0.169917035 0.49263118 -0.28187848
#> [105,] -0.132714772 0.63363118 -0.36007016
#> [106,] -0.010340372 0.64490104 -0.35816820
#> [107,]  0.192509988 0.56365816 -0.28489600
#> [108,]  0.353680730 0.47412303 -0.17488729
#> [109,]  0.432737810 0.40336372 -0.02878661
#> [110,] -0.090224045 0.37244888  0.71769159
#> [111,]  0.000000000 0.00000000  0.00000000
#> [112,] -0.031886120 0.07479893 -0.04282603
#> [113,] -0.062585869 0.14804054 -0.08476042
#> [114,] -0.113858263 0.27859466 -0.15950901
#> [115,] -0.168652039 0.48185238 -0.27574137
#> [116,] -0.129701084 0.63575371 -0.36111993
#> [117,] -0.013265008 0.64554696 -0.35878460
#> [118,]  0.187939184 0.56594061 -0.28713439
#> [119,]  0.335107293 0.48587347 -0.19248177
#> [120,]  0.432921151 0.40296875 -0.02765524
#> [121,] -0.118883480 0.37506632  0.74379911
#> [122,]  0.000000000 0.00000000  0.00000000
#> [123,] -0.031886120 0.07479893 -0.04282603
#> [124,] -0.062585869 0.14804054 -0.08476042
#> [125,] -0.115680778 0.28362245 -0.16238767
#> [126,] -0.169045005 0.48497498 -0.27751992
#> [127,] -0.134477691 0.63228862 -0.35939765
#> [128,] -0.008383944 0.64445532 -0.35774656
#> [129,]  0.191599626 0.56411346 -0.28534422
#> [130,]  0.335107293 0.48587347 -0.19248177
#> [131,]  0.432255723 0.40435862 -0.03161162
#> [132,] -0.124174302 0.37555209  0.74861003
#> attr(,"degree")
#> [1] 3
#> attr(,"knots")
#> [1] 1.040000 7.023333
#> attr(,"Boundary.knots")
#> [1]  0.00 24.65
#> attr(,"intercept")
#> [1] FALSE
#> attr(,"class")
#> [1] "ns"     "basis"  "matrix"

# Or, compute the spline basis beforehand, and then pass it to time_spline()
basis <-
  splines::bs(Theoph$Time, df = 3, Boundary.knots = c(0, max(Theoph$Time)))

time_spline(Theoph$Time, basis = basis)
#>                   1            2            3
#>   [1,] 0.0000000000 0.0000000000 0.000000e+00
#>   [2,] 0.0298119336 0.0003054501 1.043204e-06
#>   [3,] 0.0662000481 0.0015670277 1.236442e-05
#>   [4,] 0.1242030786 0.0059119187 9.380010e-05
#>   [5,] 0.2072005774 0.0184951465 5.503048e-04
#>   [6,] 0.3319803563 0.0608816592 3.721682e-03
#>   [7,] 0.3904219115 0.1018491943 8.856452e-03
#>   [8,] 0.4371571813 0.1744162874 2.319611e-02
#>   [9,] 0.4411316921 0.2559129368 4.948744e-02
#>  [10,] 0.3811318524 0.3686606585 1.188658e-01
#>  [11,] 0.0003826856 0.0333073111 9.663085e-01
#>  [12,] 0.0000000000 0.0000000000 0.000000e+00
#>  [13,] 0.0321441281 0.0003559850 1.314137e-06
#>  [14,] 0.0606440883 0.0013068763 9.387701e-06
#>  [15,] 0.1120295971 0.0047369808 6.676506e-05
#>  [16,] 0.1986875177 0.0167831075 4.725556e-04
#>  [17,] 0.3135878818 0.0518939757 2.862552e-03
#>  [18,] 0.3874492283 0.0990827879 8.446181e-03
#>  [19,] 0.4371571813 0.1744162874 2.319611e-02
#>  [20,] 0.4415111513 0.2539041765 4.867173e-02
#>  [21,] 0.3846208125 0.3648576877 1.153700e-01
#>  [22,] 0.0005962287 0.0413953039 9.580056e-01
#>  [23,] 0.0000000000 0.0000000000 0.000000e+00
#>  [24,] 0.0321441281 0.0003559850 1.314137e-06
#>  [25,] 0.0673055160 0.0016218197 1.302666e-05
#>  [26,] 0.1140770019 0.0049241871 7.085161e-05
#>  [27,] 0.2072005774 0.0184951465 5.503048e-04
#>  [28,] 0.3206694595 0.0551984519 3.167196e-03
#>  [29,] 0.3896869368 0.1011553214 8.752666e-03
#>  [30,] 0.4376507118 0.1760062874 2.359432e-02
#>  [31,] 0.4415111513 0.2539041765 4.867173e-02
#>  [32,] 0.3802478681 0.3696009278 1.197507e-01
#>  [33,] 0.0011153974 0.0561649042 9.427123e-01
#>  [34,] 0.0000000000 0.0000000000 0.000000e+00
#>  [35,] 0.0413953039 0.0005962287 2.862552e-06
#>  [36,] 0.0695107374 0.0017341556 1.442125e-05
#>  [37,] 0.1191631193 0.0054073171 8.179007e-05
#>  [38,] 0.2163649256 0.0204643558 6.451906e-04
#>  [39,] 0.3135878818 0.0518939757 2.862552e-03
#>  [40,] 0.3874492283 0.0990827879 8.446181e-03
#>  [41,] 0.4370309765 0.1740191410 2.309727e-02
#>  [42,] 0.4413620392 0.2547079714 4.899693e-02
#>  [43,] 0.3851949029 0.3642174378 1.147941e-01
#>  [44,] 0.0000000000 0.0000000000 1.000000e+00
#>  [45,] 0.0000000000 0.0000000000 0.000000e+00
#>  [46,] 0.0356278545 0.0004389469 1.802657e-06
#>  [47,] 0.0606440883 0.0013068763 9.387701e-06
#>  [48,] 0.1120295971 0.0047369808 6.676506e-05
#>  [49,] 0.2072005774 0.0184951465 5.503048e-04
#>  [50,] 0.3135878818 0.0518939757 2.862552e-03
#>  [51,] 0.3874492283 0.0990827879 8.446181e-03
#>  [52,] 0.4370309765 0.1740191410 2.309727e-02
#>  [53,] 0.4407300502 0.2579191934 5.031221e-02
#>  [54,] 0.3846208125 0.3648576877 1.153700e-01
#>  [55,] 0.0004389469 0.0356278545 9.639314e-01
#>  [56,] 0.0000000000 0.0000000000 0.000000e+00
#>  [57,] 0.0321441281 0.0003559850 1.314137e-06
#>  [58,] 0.0673055160 0.0016218197 1.302666e-05
#>  [59,] 0.1272049609 0.0062249236 1.015413e-04
#>  [60,] 0.2080423367 0.0186704661 5.585182e-04
#>  [61,] 0.3177458690 0.0538118004 3.037763e-03
#>  [62,] 0.3866923653 0.0983950039 8.345632e-03
#>  [63,] 0.4367751719 0.1732252807 2.290041e-02
#>  [64,] 0.4396765032 0.2627230952 5.232895e-02
#>  [65,] 0.3817185855 0.3680314649 1.182784e-01
#>  [66,] 0.0030573055 0.0911459210 9.057626e-01
#>  [67,] 0.0000000000 0.0000000000 0.000000e+00
#>  [68,] 0.0298119336 0.0003054501 1.043204e-06
#>  [69,] 0.0584083259 0.0012092821 8.345632e-06
#>  [70,] 0.1140770019 0.0049241871 7.085161e-05
#>  [71,] 0.2072005774 0.0184951465 5.503048e-04
#>  [72,] 0.3123859149 0.0513511093 2.813759e-03
#>  [73,] 0.3866923653 0.0983950039 8.345632e-03
#>  [74,] 0.4365148327 0.1724320052 2.270469e-02
#>  [75,] 0.4415111513 0.2539041765 4.867173e-02
#>  [76,] 0.3831762837 0.3664503348 1.168182e-01
#>  [77,] 0.0008969775 0.0505227774 9.485749e-01
#>  [78,] 0.0000000000 0.0000000000 0.000000e+00
#>  [79,] 0.0298119336 0.0003054501 1.043204e-06
#>  [80,] 0.0606440883 0.0013068763 9.387701e-06
#>  [81,] 0.1099747733 0.0045532437 6.283874e-05
#>  [82,] 0.2072005774 0.0184951465 5.503048e-04
#>  [83,] 0.3153791781 0.0527125236 2.936793e-03
#>  [84,] 0.3885742386 0.1001173421 8.598513e-03
#>  [85,] 0.4385838373 0.1791928250 2.440436e-02
#>  [86,] 0.4409736889 0.2567157483 4.981626e-02
#>  [87,] 0.3817185855 0.3680314649 1.182784e-01
#>  [88,] 0.0013570615 0.0617591003 9.368739e-01
#>  [89,] 0.0000000000 0.0000000000 0.000000e+00
#>  [90,] 0.0356278545 0.0004389469 1.802657e-06
#>  [91,] 0.0728043015 0.0019095216 1.669440e-05
#>  [92,] 0.1171342197 0.0052114801 7.728890e-05
#>  [93,] 0.2072005774 0.0184951465 5.503048e-04
#>  [94,] 0.3153791781 0.0527125236 2.936793e-03
#>  [95,] 0.3874492283 0.0990827879 8.446181e-03
#>  [96,] 0.4388059383 0.1799907653 2.460972e-02
#>  [97,] 0.4428041568 0.2458471028 4.549852e-02
#>  [98,] 0.3956849176 0.3517199267 1.042133e-01
#>  [99,] 0.0002368314 0.0262990523 9.734634e-01
#> [100,] 0.0000000000 0.0000000000 0.000000e+00
#> [101,] 0.0436887453 0.0006657675 3.381850e-06
#> [102,] 0.0879487870 0.0028358696 3.048045e-05
#> [103,] 0.1140770019 0.0049241871 7.085161e-05
#> [104,] 0.2097206636 0.0190233345 5.751893e-04
#> [105,] 0.3165656201 0.0532610404 2.986994e-03
#> [106,] 0.3885742386 0.1001173421 8.598513e-03
#> [107,] 0.4377712783 0.1764041349 2.369458e-02
#> [108,] 0.4380779522 0.2691009294 5.510078e-02
#> [109,] 0.3817185855 0.3680314649 1.182784e-01
#> [110,] 0.0042841635 0.1068786061 8.887800e-01
#> [111,] 0.0000000000 0.0000000000 0.000000e+00
#> [112,] 0.0298119336 0.0003054501 1.043204e-06
#> [113,] 0.0584083259 0.0012092821 8.345632e-06
#> [114,] 0.1099747733 0.0045532437 6.283874e-05
#> [115,] 0.2038162066 0.0178013273 5.182566e-04
#> [116,] 0.3195046513 0.0546421256 3.114991e-03
#> [117,] 0.3874492283 0.0990827879 8.446181e-03
#> [118,] 0.4371571813 0.1744162874 2.319611e-02
#> [119,] 0.4412861461 0.2551097247 4.916007e-02
#> [120,] 0.3811318524 0.3686606585 1.188658e-01
#> [121,] 0.0015670277 0.0662000481 9.322206e-01
#> [122,] 0.0000000000 0.0000000000 0.000000e+00
#> [123,] 0.0298119336 0.0003054501 1.043204e-06
#> [124,] 0.0584083259 0.0012092821 8.345632e-06
#> [125,] 0.1120295971 0.0047369808 6.676506e-05
#> [126,] 0.2055118611 0.0181467427 5.341205e-04
#> [127,] 0.3147836316 0.0524391095 2.911905e-03
#> [128,] 0.3893174034 0.1008089497 8.701079e-03
#> [129,] 0.4376507118 0.1760062874 2.359432e-02
#> [130,] 0.4412861461 0.2551097247 4.916007e-02
#> [131,] 0.3831762837 0.3664503348 1.168182e-01
#> [132,] 0.0012092821 0.0584083259 9.403740e-01
#> attr(,"degree")
#> [1] 3
#> attr(,"knots")
#> numeric(0)
#> attr(,"Boundary.knots")
#> [1]  0.00 24.65
#> attr(,"intercept")
#> [1] FALSE
#> attr(,"class")
#> [1] "splinetrials_time_spline" "bs"                      
#> [3] "basis"                    "matrix"                  
#> attr(,"basis")
#>                   1            2            3
#>   [1,] 0.0000000000 0.0000000000 0.000000e+00
#>   [2,] 0.0298119336 0.0003054501 1.043204e-06
#>   [3,] 0.0662000481 0.0015670277 1.236442e-05
#>   [4,] 0.1242030786 0.0059119187 9.380010e-05
#>   [5,] 0.2072005774 0.0184951465 5.503048e-04
#>   [6,] 0.3319803563 0.0608816592 3.721682e-03
#>   [7,] 0.3904219115 0.1018491943 8.856452e-03
#>   [8,] 0.4371571813 0.1744162874 2.319611e-02
#>   [9,] 0.4411316921 0.2559129368 4.948744e-02
#>  [10,] 0.3811318524 0.3686606585 1.188658e-01
#>  [11,] 0.0003826856 0.0333073111 9.663085e-01
#>  [12,] 0.0000000000 0.0000000000 0.000000e+00
#>  [13,] 0.0321441281 0.0003559850 1.314137e-06
#>  [14,] 0.0606440883 0.0013068763 9.387701e-06
#>  [15,] 0.1120295971 0.0047369808 6.676506e-05
#>  [16,] 0.1986875177 0.0167831075 4.725556e-04
#>  [17,] 0.3135878818 0.0518939757 2.862552e-03
#>  [18,] 0.3874492283 0.0990827879 8.446181e-03
#>  [19,] 0.4371571813 0.1744162874 2.319611e-02
#>  [20,] 0.4415111513 0.2539041765 4.867173e-02
#>  [21,] 0.3846208125 0.3648576877 1.153700e-01
#>  [22,] 0.0005962287 0.0413953039 9.580056e-01
#>  [23,] 0.0000000000 0.0000000000 0.000000e+00
#>  [24,] 0.0321441281 0.0003559850 1.314137e-06
#>  [25,] 0.0673055160 0.0016218197 1.302666e-05
#>  [26,] 0.1140770019 0.0049241871 7.085161e-05
#>  [27,] 0.2072005774 0.0184951465 5.503048e-04
#>  [28,] 0.3206694595 0.0551984519 3.167196e-03
#>  [29,] 0.3896869368 0.1011553214 8.752666e-03
#>  [30,] 0.4376507118 0.1760062874 2.359432e-02
#>  [31,] 0.4415111513 0.2539041765 4.867173e-02
#>  [32,] 0.3802478681 0.3696009278 1.197507e-01
#>  [33,] 0.0011153974 0.0561649042 9.427123e-01
#>  [34,] 0.0000000000 0.0000000000 0.000000e+00
#>  [35,] 0.0413953039 0.0005962287 2.862552e-06
#>  [36,] 0.0695107374 0.0017341556 1.442125e-05
#>  [37,] 0.1191631193 0.0054073171 8.179007e-05
#>  [38,] 0.2163649256 0.0204643558 6.451906e-04
#>  [39,] 0.3135878818 0.0518939757 2.862552e-03
#>  [40,] 0.3874492283 0.0990827879 8.446181e-03
#>  [41,] 0.4370309765 0.1740191410 2.309727e-02
#>  [42,] 0.4413620392 0.2547079714 4.899693e-02
#>  [43,] 0.3851949029 0.3642174378 1.147941e-01
#>  [44,] 0.0000000000 0.0000000000 1.000000e+00
#>  [45,] 0.0000000000 0.0000000000 0.000000e+00
#>  [46,] 0.0356278545 0.0004389469 1.802657e-06
#>  [47,] 0.0606440883 0.0013068763 9.387701e-06
#>  [48,] 0.1120295971 0.0047369808 6.676506e-05
#>  [49,] 0.2072005774 0.0184951465 5.503048e-04
#>  [50,] 0.3135878818 0.0518939757 2.862552e-03
#>  [51,] 0.3874492283 0.0990827879 8.446181e-03
#>  [52,] 0.4370309765 0.1740191410 2.309727e-02
#>  [53,] 0.4407300502 0.2579191934 5.031221e-02
#>  [54,] 0.3846208125 0.3648576877 1.153700e-01
#>  [55,] 0.0004389469 0.0356278545 9.639314e-01
#>  [56,] 0.0000000000 0.0000000000 0.000000e+00
#>  [57,] 0.0321441281 0.0003559850 1.314137e-06
#>  [58,] 0.0673055160 0.0016218197 1.302666e-05
#>  [59,] 0.1272049609 0.0062249236 1.015413e-04
#>  [60,] 0.2080423367 0.0186704661 5.585182e-04
#>  [61,] 0.3177458690 0.0538118004 3.037763e-03
#>  [62,] 0.3866923653 0.0983950039 8.345632e-03
#>  [63,] 0.4367751719 0.1732252807 2.290041e-02
#>  [64,] 0.4396765032 0.2627230952 5.232895e-02
#>  [65,] 0.3817185855 0.3680314649 1.182784e-01
#>  [66,] 0.0030573055 0.0911459210 9.057626e-01
#>  [67,] 0.0000000000 0.0000000000 0.000000e+00
#>  [68,] 0.0298119336 0.0003054501 1.043204e-06
#>  [69,] 0.0584083259 0.0012092821 8.345632e-06
#>  [70,] 0.1140770019 0.0049241871 7.085161e-05
#>  [71,] 0.2072005774 0.0184951465 5.503048e-04
#>  [72,] 0.3123859149 0.0513511093 2.813759e-03
#>  [73,] 0.3866923653 0.0983950039 8.345632e-03
#>  [74,] 0.4365148327 0.1724320052 2.270469e-02
#>  [75,] 0.4415111513 0.2539041765 4.867173e-02
#>  [76,] 0.3831762837 0.3664503348 1.168182e-01
#>  [77,] 0.0008969775 0.0505227774 9.485749e-01
#>  [78,] 0.0000000000 0.0000000000 0.000000e+00
#>  [79,] 0.0298119336 0.0003054501 1.043204e-06
#>  [80,] 0.0606440883 0.0013068763 9.387701e-06
#>  [81,] 0.1099747733 0.0045532437 6.283874e-05
#>  [82,] 0.2072005774 0.0184951465 5.503048e-04
#>  [83,] 0.3153791781 0.0527125236 2.936793e-03
#>  [84,] 0.3885742386 0.1001173421 8.598513e-03
#>  [85,] 0.4385838373 0.1791928250 2.440436e-02
#>  [86,] 0.4409736889 0.2567157483 4.981626e-02
#>  [87,] 0.3817185855 0.3680314649 1.182784e-01
#>  [88,] 0.0013570615 0.0617591003 9.368739e-01
#>  [89,] 0.0000000000 0.0000000000 0.000000e+00
#>  [90,] 0.0356278545 0.0004389469 1.802657e-06
#>  [91,] 0.0728043015 0.0019095216 1.669440e-05
#>  [92,] 0.1171342197 0.0052114801 7.728890e-05
#>  [93,] 0.2072005774 0.0184951465 5.503048e-04
#>  [94,] 0.3153791781 0.0527125236 2.936793e-03
#>  [95,] 0.3874492283 0.0990827879 8.446181e-03
#>  [96,] 0.4388059383 0.1799907653 2.460972e-02
#>  [97,] 0.4428041568 0.2458471028 4.549852e-02
#>  [98,] 0.3956849176 0.3517199267 1.042133e-01
#>  [99,] 0.0002368314 0.0262990523 9.734634e-01
#> [100,] 0.0000000000 0.0000000000 0.000000e+00
#> [101,] 0.0436887453 0.0006657675 3.381850e-06
#> [102,] 0.0879487870 0.0028358696 3.048045e-05
#> [103,] 0.1140770019 0.0049241871 7.085161e-05
#> [104,] 0.2097206636 0.0190233345 5.751893e-04
#> [105,] 0.3165656201 0.0532610404 2.986994e-03
#> [106,] 0.3885742386 0.1001173421 8.598513e-03
#> [107,] 0.4377712783 0.1764041349 2.369458e-02
#> [108,] 0.4380779522 0.2691009294 5.510078e-02
#> [109,] 0.3817185855 0.3680314649 1.182784e-01
#> [110,] 0.0042841635 0.1068786061 8.887800e-01
#> [111,] 0.0000000000 0.0000000000 0.000000e+00
#> [112,] 0.0298119336 0.0003054501 1.043204e-06
#> [113,] 0.0584083259 0.0012092821 8.345632e-06
#> [114,] 0.1099747733 0.0045532437 6.283874e-05
#> [115,] 0.2038162066 0.0178013273 5.182566e-04
#> [116,] 0.3195046513 0.0546421256 3.114991e-03
#> [117,] 0.3874492283 0.0990827879 8.446181e-03
#> [118,] 0.4371571813 0.1744162874 2.319611e-02
#> [119,] 0.4412861461 0.2551097247 4.916007e-02
#> [120,] 0.3811318524 0.3686606585 1.188658e-01
#> [121,] 0.0015670277 0.0662000481 9.322206e-01
#> [122,] 0.0000000000 0.0000000000 0.000000e+00
#> [123,] 0.0298119336 0.0003054501 1.043204e-06
#> [124,] 0.0584083259 0.0012092821 8.345632e-06
#> [125,] 0.1120295971 0.0047369808 6.676506e-05
#> [126,] 0.2055118611 0.0181467427 5.341205e-04
#> [127,] 0.3147836316 0.0524391095 2.911905e-03
#> [128,] 0.3893174034 0.1008089497 8.701079e-03
#> [129,] 0.4376507118 0.1760062874 2.359432e-02
#> [130,] 0.4412861461 0.2551097247 4.916007e-02
#> [131,] 0.3831762837 0.3664503348 1.168182e-01
#> [132,] 0.0012092821 0.0584083259 9.403740e-01
#> attr(,"degree")
#> [1] 3
#> attr(,"knots")
#> numeric(0)
#> attr(,"Boundary.knots")
#> [1]  0.00 24.65
#> attr(,"intercept")
#> [1] FALSE
#> attr(,"class")
#> [1] "bs"     "basis"  "matrix"