splinetrials_analysis object

ncs_analysis() returns an object of class splinetrials_analysis: a 32-column tibble with one row per unique combination of data[[arm]] and data[[time_scheduled_label]] (see the arguments of ncs_analysis()).

Columns

arm: values of data[[arm]].

time: values of data[[time_scheduled_label]].

n: number of times the combination appears in data.

est: mean of data[[response]].

sd: standard deviation of data[[response]].

se: standard error of data[[response]] (i.e., sd / sqrt(n)).

lower: lower bound of confidence interval.

upper: upper bound of confidence interval.

response_est: estimated marginal mean.

response_se: standard error of response_est.

response_df: degrees of freedom used for calculating the confidence interval for response_est.

response_lower: lower bound of confidence interval for response_est.

response_upper: upper bound of confidence interval for response_est.

change_est: estimated change from baseline.

change_se: standard error of change_est.

change_df: degrees of freedom used for calculating the confidence interval for and testing the significance of change_est.

change_lower: lower bound of confidence interval for change_est.

change_upper: upper bound of confidence interval for change_est.

change_test_statistic: test statistic measuring the significance of change_est.

change_p_value: p-value for the significance of change_est.

diff_est: treatment effect.

diff_se: standard error of diff_est.

diff_df: degrees of freedom used for calculating the confidence interval for and testing the significance of diff_est.

diff_lower: lower bound of confidence interval for diff_est.

diff_upper: upper bound of confidence interval for diff_est.

diff_test_statistic: test statistic measuring the significance of diff_est.

diff_p_value: p-value for the significance of diff_est.

percent_slowing_est: estimated percent slowing.

percent_slowing_lower: lower bound of confidence interval for percent_slowing_est.

percent_slowing_upper: upper bound of confidence interval for percent_slowing_est.

correlation: the covariance structure of the analysis model. This is the same value repeated for each row.

optimizer: invariably mmrm+tmb to indicate that mmrm::mmrm() (which uses the TMB package) was used to fit the model.

Optional analysis_model attribute

If ncs_analysis() had return_models = TRUE, then the analysis model, an mmrm object, will be included as the analysis_model attribute.

See also